+++ advtraj/__init__.py
@@ -1,1 +1,1 @@
-from .integrate import integrate_trajectories  # noqa
+from advtraj.integrate import integrate_trajectories  # noqa