+++ advtraj/integrate/__init__.py
@@ -1,10 +1,10 @@
 """
 Main routines for integration
 """
-from .backward import backward as integrate_backward
-from .forward import forward as integrate_forward
+from advtraj.integrate.backward import backward as integrate_backward
+from advtraj.integrate.forward import forward as integrate_forward
 
-POSITION_COORD_NAMES = ["x", "y", "z", "time"]
+POSITION_COORD_NAMES = ["x", "y", "z"]
 
 
 def _validate_position_scalars(ds, xy_periodic=False):
@@ -51,11 +51,30 @@ def _validate_starting_points(ds):
             f" {', '.join(missing_vars)}"
         )
 
+    if 'time' not in ds.coords:
+        raise Exception(
+            "The starting position dataset is missing the time cooord."
+        )
 
-def _integrate_single_trajectory(ds_starting_point, ds_position_scalars, da_times):
+    coords = [ c for c in ds[POSITION_COORD_NAMES[0]].coords
+                    if 'time' not in c]
+
+    if len(coords) == 0:
+        raise Exception(
+            "The starting position dataset is missing the trajectory coord."
+        )
+
+
+def _integrate_single_trajectory(ds_starting_point,
+                                 ds_position_scalars,
+                                 da_times,
+                                 interp_order = 1
+):
     """
+    Integrate single trajectory forwards and back.
+
     Integrate a single trajectory both forward and backward from
-    `ds_starting_point` usnig `ds_position_scalars` and times in `da_times`
+    `ds_starting_point` using `ds_position_scalars` and times in `da_times`
     """
     da_times_backward = da_times.sel(time=slice(None, ds_starting_point.time))
     da_times_forward = da_times.sel(time=slice(ds_starting_point.time, None)).isel(
@@ -66,9 +85,8 @@ def _integrate_single_trajectory(ds_starting_point, ds_position_scalars, da_time
         ds_position_scalars=ds_position_scalars,
         ds_starting_point=ds_starting_point,
         da_times=da_times_backward,
+        interp_order=interp_order,
     )
-
-    interp_order = 1
 
     ds_traj = integrate_forward(
         ds_position_scalars=ds_position_scalars,
@@ -79,11 +97,48 @@ def _integrate_single_trajectory(ds_starting_point, ds_position_scalars, da_time
 
     return ds_traj
 
-
-def integrate_trajectories(
-    ds_position_scalars, ds_starting_points, times="position_scalars", xy_periodic=True
+def _integrate_trajectories(ds_starting_point,
+                                     ds_position_scalars,
+                                     da_times,
+                                     interp_order = 1
 ):
     """
+    Integrate multiple trajectories forwards and back.
+
+    Integrate a trajectories both forward and backward from
+    `ds_starting_point` using `ds_position_scalars` and times in `da_times`
+    """
+    ref_time = ds_starting_point.time
+    da_times_backward = da_times.sel(time=slice(None, ref_time))
+    da_times_forward = da_times.sel(time=slice(ref_time, None)).isel(
+        time=slice(1, None)
+    )
+
+    ds_traj_backward = integrate_backward(
+        ds_position_scalars=ds_position_scalars,
+        ds_starting_point=ds_starting_point,
+        da_times=da_times_backward,
+        interp_order=interp_order,
+    )
+
+    ds_traj = integrate_forward(
+        ds_position_scalars=ds_position_scalars,
+        ds_back_trajectory=ds_traj_backward,
+        da_times=da_times_forward,
+        interp_order=interp_order,
+    )
+
+    return ds_traj
+
+def integrate_trajectories(ds_position_scalars,
+                           ds_starting_points,
+                           times="position_scalars",
+                           xy_periodic=True,
+                           interp_order = 1
+):
+    """
+    Integrate trajectories forwards and back.
+
     Using "position scalars" `ds_position_scalars` integrate trajectories from
     starting points in `ds_starting_points` to times as in `times`
     """
@@ -97,24 +152,12 @@ def integrate_trajectories(
 
     # all coordinates that are defined for the starting position variables will
     # be treated as if they represent different trajectories
-    dims = ds_starting_points[POSITION_COORD_NAMES[0]].dims
+    ds_traj = _integrate_trajectories(
+        ds_position_scalars=ds_position_scalars,
+        ds_starting_point=ds_starting_points,
+        da_times=da_times,
+        interp_order=interp_order,
+    )
 
-    if len(dims) > 0:
-        ds_traj = (
-            ds_starting_points.stack(trajectory=dims)
-            .groupby("trajectory")
-            .apply(
-                _integrate_single_trajectory,
-                ds_position_scalars=ds_position_scalars,
-                da_times=da_times,
-            )
-            .unstack("trajectory")
-        )
-    else:
-        ds_traj = _integrate_single_trajectory(
-            ds_position_scalars=ds_position_scalars,
-            ds_starting_point=ds_starting_points,
-            da_times=da_times,
-        )
 
     return ds_traj