+++ advtraj/lib/__init__.py
@@ -1,1 +1,0 @@
-from .fast_interp import fast_interp  # noqa