+++ advtraj/utils/interpolation.py
@@ -6,7 +6,8 @@ with (optional) cyclic boundary conditions
 import numpy as np
 import xarray as xr
 
-from ..lib import fast_interp
+#from ..lib import fast_interp # this would be better as a separate install.
+import fast_interp
 from .grid import find_grid_spacing
 
 
@@ -51,18 +52,19 @@ def interpolate_3d_field(da, ds_positions, interp_order=1, cyclic_boundaries=[])
     have cyclic boundaries, e.g. (`cyclic_boundaries = 'xy'` or
     `cyclic_boundaries = ['x', 'y']`)
     """
-    dx, dy, dz = find_grid_spacing(da, coords="xyz")
+#    dx, dy, dz = find_grid_spacing(da, coords="xyz")
 
     c_min = np.array([da[c].min().values for c in da.dims])
     c_max = np.array([da[c].max().values for c in da.dims])
-    dX = np.array([dx, dy, dz])
+#    dX = np.array([dx, dy, dz])
+    dX = np.array([da[c].attrs[f'd{c}'] for c in da.dims])
     periodicity = [c in cyclic_boundaries for c in da.dims]
 
     fn = fast_interp.interp3d(
         a=c_min, b=c_max, h=dX, f=da.values, k=interp_order, p=periodicity
     )
 
-    vals = fn(*[ds_positions[c] for c in da.dims])
+    vals = fn(*[ds_positions[c].values for c in da.dims])
 
     da_interpolated = xr.DataArray(
         vals,
@@ -74,11 +76,35 @@ def interpolate_3d_field(da, ds_positions, interp_order=1, cyclic_boundaries=[])
 
     return da_interpolated
 
-
-def interpolate_3d_fields(ds, ds_positions, interp_order=1, cyclic_boundaries=None):
+def interpolate_from_interpolator(v, ds_positions, fn):
     """
-    Perform interpolation of xr.DataArray `da` at positions given by data
-    variables in `ds_posisions` with interpolation order `interp_order`. Cyclic
+    Perform interpolation of variable named 'v' at positions given by data
+    variables in `ds_positions` using interpolator fn.
+    """
+
+    vals = fn(*[ds_positions[c].values for c in 'xyz'])
+
+    da_interpolated = xr.DataArray(
+        vals,
+        dims=ds_positions.dims,
+        coords=ds_positions.coords,
+#        attrs=da.attrs, # do we need these?
+        name=v,
+    )
+
+    return da_interpolated
+
+def interpolate_3d_fields(ds,
+                          ds_positions,
+                          interpolator=None,
+                          interp_order=1,
+                          cyclic_boundaries=None):
+    """
+    Perform interpolation of xr.DataArray `ds` at positions given by data
+    variables in `ds_positions`.
+    If interpolator provided, look in this for pre-generated fast_inter
+    interpolator for each variable.
+    Otherwise, interpolate with interpolation order `interp_order`. Cyclic
     boundary conditions are used by providing a `list` of the coordinates which
     have cyclic boundaries, e.g. (`cyclic_boundaries = 'xy'` or
     `cyclic_boundaries = ['x', 'y']`)
@@ -86,14 +112,58 @@ def interpolate_3d_fields(ds, ds_positions, interp_order=1, cyclic_boundaries=No
     dataarrays = []
 
     for v in ds.data_vars:
-        da_interpolated = interpolate_3d_field(
-            da=ds[v],
-            ds_positions=ds_positions,
-            interp_order=interp_order,
-            cyclic_boundaries=cyclic_boundaries,
-        )
+        if interpolator is not None and v in interpolator:
+
+            da_interpolated = interpolate_from_interpolator(
+                                v, ds_positions, interpolator[v])
+
+        else:
+
+            da_interpolated = interpolate_3d_field(
+                da=ds[v],
+                ds_positions=ds_positions,
+                interp_order=interp_order,
+                cyclic_boundaries=cyclic_boundaries,
+            )
         dataarrays.append(da_interpolated)
 
     ds_interpolated = xr.merge(dataarrays)
     ds_interpolated.attrs.update(ds.attrs)
     return ds_interpolated
+
+def gen_interpolator_3d_field(da, interp_order=1, cyclic_boundaries=None):
+    """
+    Generate fast_interp interpolators for xr.DataArray `da` at positions with
+    interpolation order `interp_order`. Cyclic
+    boundary conditions are used by providing a `list` of the coordinates which
+    have cyclic boundaries, e.g. (`cyclic_boundaries = 'xy'` or
+    `cyclic_boundaries = ['x', 'y']`)
+    """
+    c_min = np.array([da[c].min().values for c in da.dims])
+    c_max = np.array([da[c].max().values for c in da.dims])
+    dX = np.array([da[c].attrs[f'd{c}'] for c in da.dims])
+    periodicity = [c in cyclic_boundaries for c in da.dims]
+
+    fn = fast_interp.interp3d(
+        a=c_min, b=c_max, h=dX, f=da.values, k=interp_order, p=periodicity
+    )
+    return fn
+
+
+def gen_interpolator_3d_fields(ds, interp_order=1, cyclic_boundaries=None):
+    """
+    Generate fast_interp interpolators for xr.DataArray `da` at positions with
+    interpolation order `interp_order`. Cyclic
+    boundary conditions are used by providing a `list` of the coordinates which
+    have cyclic boundaries, e.g. (`cyclic_boundaries = 'xy'` or
+    `cyclic_boundaries = ['x', 'y']`)
+    """
+    interpolators = {}
+    for v in ds.data_vars:
+        interpolators[v] = gen_interpolator_3d_field(
+            da=ds[v],
+            interp_order=interp_order,
+            cyclic_boundaries=cyclic_boundaries,
+        )
+
+    return interpolators