+++ scripts/run_tests.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Feb 25 18:37:59 2022
+
+@author: paclk
+"""
+
+from test_synthetic_data import *
+from test_grid_utils import *
+test_cyclic_coord_wrapping("")
+# test_stationary_trajectory()
+# test_linear_trajectory_x_direction()