+++ tests/utils.py
@@ -62,7 +62,7 @@ def orange(*args, **kwargs):
     return cust_range(*args, **kwargs, include=[True, False])
 
 
-def create_uniform_grid(dL, L, cell_centered=True):
+def create_uniform_grid(dL, L, grid_style = 'monc'):
     """
     Create a grid with uniform resolution in x,y and z with domain spanning
     [0,0,0] to `L` with grid resolution `dL`
@@ -70,14 +70,18 @@ def create_uniform_grid(dL, L, cell_centered=True):
     Lx, Ly, Lz = L
     dx, dy, dz = dL
 
-    if cell_centered:
+    if grid_style == 'cell_centered':
         # create cell-center positions
         x_ = crange(dx / 2.0, Lx - dx / 2.0, dx)
         y_ = crange(dy / 2.0, Ly - dy / 2.0, dy)
         z_ = crange(dz / 2.0, Lz - dz / 2.0, dz)
+    elif grid_style == 'monc':
+        x_ = np.arange(round(Lx/dx)) * dx
+        y_ = np.arange(round(Ly/dy)) * dy
+        z_ = (np.arange(round(Lz/dz)+ 1) - 0.5) * dz
     else:
-        x_ = crange(0.0, Lx, dx)
-        y_ = crange(0.0, Ly, dy)
+        x_ = crange(0.0, Lx - dx, dx)
+        y_ = crange(0.0, Ly - dy, dy)
         z_ = crange(0.0, Lz, dz)
 
     ds = xr.Dataset(coords=dict(x=x_, y=y_, z=z_))
@@ -102,10 +106,16 @@ def create_initial_dataset(dL, L, xy_periodic=True):
     dx, dy, dz = dL
     ds_grid = create_uniform_grid(dL=dL, L=L)
 
-    ds_grid["xy_periodic"] = xy_periodic
+    ds_grid.attrs["xy_periodic"] = xy_periodic
 
     ds = init_position_scalars(ds=ds_grid)
-    ds["time"] = np.datetime64("2020-01-01T00:00")
+#    ds["time"] = np.datetime64("2020-01-01T00:00")
+    ds = ds.assign_coords(time=np.datetime64("2020-01-01T00:00"))
+
+    for i, c in enumerate("xyz"):
+
+        ds[c].attrs[f"d{c}"] = dL[i]
+        ds[c].attrs[f"L{c}"] = L[i]
 
     return ds